// repos/sbs-draft-web/utils/types/types.ts
export type PlayerStateInfo = {
    playerId: string
    displayName: string
    team: string
    position: string
    ownerAddress: string
    pickNum: number
    round: number
}

export type RealTimeDraftInfo = {
    currentDrafter: string
    pickNumber: number
    roundNum: number
    pickInRound: number
    pickEndTime: number        // Unix timestamp in seconds
    pickLength: number         // Duration in seconds
    draftStartTime: number     // Unix timestamp in seconds when draft starts
    lastPick: PlayerStateInfo | null
    isDraftComplete: boolean
    isDraftClosed: boolean
}